Several major changes - first part

+++ VolleyManagement.UnitTests/Services/PlayerService/PlayerBuilder.cs
@@ -70,7 +70,7 @@
         /// </summary>
         /// <param name="birthYear">Test player birth year</param>
         /// <returns>Player builder object</returns>
-        public PlayerBuilder WithBirthYear(int? birthYear)
+        public PlayerBuilder WithBirthYear(short? birthYear)
         {
             _player.BirthYear = birthYear;
             return this;
@@ -81,7 +81,7 @@
         /// </summary>
         /// <param name="height">Test player height</param>
         /// <returns>Player builder object</returns>
-        public PlayerBuilder WithHeight(int? height)
+        public PlayerBuilder WithHeight(short? height)
         {
             _player.Height = height;
             return this;
@@ -92,7 +92,7 @@
         /// </summary>
         /// <param name="weight">Test player weight</param>
         /// <returns>Player builder object</returns>
-        public PlayerBuilder WithWeight(int? weight)
+        public PlayerBuilder WithWeight(short? weight)
         {
             _player.Weight = weight;
             return this;
